requirements.txt
psutil
PySide6-Essentials
